requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
alembic>=1.13.1
httpx>=0.27.0
beautifulsoup4>=4.12.3
dnspython>=2.6.1
pydantic>=2.9.2
pydantic-settings>=2.2.1
